Make tokenizer be permissive
It now should always succeed meaning that any errors must be
thrown in the parser. This allows us to use to the same tokenizer
for both normal line parsing and for colorizing output with reline.

At the same time, I updated the error messages to be much more
descriptive. They now point to the exact location where the error
occurred during parsing.