Add support for option completions and hints
This adds support for options completions and hints.

--Overview--

Options are parsed from the help page of each individual Git command.
This is done lazily the first time that an option needs to be checked
so it shouldn't end up being a performance concern. Once the options
have been parsed from the help page they are cached for the life of
the program. These options are used both for completions and hints.

--Completions--

Completions are given based on the options matching the same long prefix.
No completions are given for short options like `-a`.

Input:
git diff --out

Completions:
--output
--output-indicator-new
--output-indicator-old
--output-indicator-context

Note: Completions are not shown if the option is added after the end-of-options
token which is `--` since that indicates that options should no longer be parsed.

--Hint

Hints are shown in gray after an exact match of an option. These are shown for
both short and long options. If there are multiple hints parsed from the help
page, the first non-empty one is chosen and shown here.

Input:
git diff --stat

Hint:
[=<width>[,<name-width>[,<count>]]]

Note: Hints are not shown if the option is added after the end-of-options
token which is `--` since that indicates that options should no longer be parsed.