Add filepath completions
These are activated when a string starting with ./ is written and
is not a command. The goal is to give a more flexible mechanism
for completions outside of guessing which things might be files.

It also changes how we interact with outside processes to avoid
reading everything into memory all at once. This should be more
effective for very large Git repositories.