Make Git.commits test deterministic
If run in isolation, this test works fine. When run with other
tests it is not deterministic because the @commit_hash_to_title
instance variable might have already been cached from a previous
test. By removing this expectation the rest of the test should
be correct since every call to `Git.commits` actually overrides
that cached instance variable.